pyproject.toml
[tool.poetry]
name = "v2django"
version = "0.1.0"
description = ""
authors = ["anonopiran <anonopiran@proton.me>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
influxdb-client = {extras = ["ciso"], version = "^1.34.0"}
django-environ = "^0.9.0"
drf-spectacular = "^0.24.2"
yarl = "^1.8.1"
psycopg2 = "^2.9.5"
requests = "^2.28.1"
gunicorn = "^20.1.0"
humanize = "^4.4.0"
celery = {extras = ["librabbitmq", "django"], version = "^5.2.7"}
django-tracking-model = "^0.1.5"
django-fsm = "^2.8.1"
django-filter = "^22.1"
pika = "^1.3.1"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
Flake8-pyproject = "^1.1.0"
black = "^22.10.0"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
force-exclude = 'V2RayMan/v2ray_proto'

[tool.flake8]
ignore = ['E203', 'E266', 'E501', 'W503', 'F403', 'F401']
max-line-length = 79
max-complexity = 18
select = ['B', 'C', 'E', 'F', 'W', 'T4', 'B9']
exclude = 'V2RayMan/v2ray_proto/*'

[tool.isort]
profile = "black"
